build: Changing makefile to exclude msp430.h from cppcheck
Cppcheck is slow checking all of the ifdef's, which msp430.h contains many of.
This slows down cppcheck because it checks all the ifdef combinations.
Because this file is provided by TI, there is no need to check this file,
so we are excluding it from the cppcheck. Suppress the error about missing
system includes (ignores standard includes <>) and supress errors about unmatched supressions and unused functions.

Also adding -f flag to make "clean" to return ok even when there is no
"build" directory.