feature(io_drivers): Adding IO drivers to configure the pins
Adding drivers to config IO pins (currently only for MSP430 Launchpad)
Drivers to: select pin function
	    set pin direction
	    set pin output high/low
	    enable/disable pullup/pulldown resistor
	    Set pullup/pulldown for input pins
The pinmapping for the MSP430 are represented with enums. Other enums to define
the functions for the drivers mentioned above were also created. There are 8 ports
and 8 pins for each port, so the IO mapping enum uses xxx_yyy as the masks for port/pin.
This acts as a Hardware Abstraction Layer (HAL) to configure the pins instead of having
to directly access the registers every time.