build: Adding ability to detect HW is Launchpad or PCB based on GPIO pullup.
GPIO P3.7 set to input and measured in io_init() function to check if pin
is pulled-up or not externally. PCB will have a pullup resistor to set this
pin high. Launchpad will not have a pullup resistor connected to the pin, so
it will be low. When flashing the program to device, the program will hang in
a while loop if the wrong software is flashed to the wrong HW.