feat(LED assert): Adding assert for test LED and HW detection.
HW detection specified in make cmd with HW=SUMOBOT or HW=LAUNCHPAD. If specifying wrong
HW, then an assert is triggered, hanging the device in an endless while loop. Tested with
blinking IO_TEST_LED, which will be used later when an assert happens, endlessly blink the
LED until device is restarted.