feat(assert): Adding static assert to check the io_generic_enum is < 1 byte.
Adding a compiler flag (-fshort-enums) to make enums as small as possible. Because
all of the values inside the enum fit within 0-255 (8 bits), the enum can be compressed
to a size of 1 byte, saving memory space. Static assert for this also prevents the code
from compiling if the enum is too big.