feat(io_interrupts): Adding interrupts on IO pins and increasing clock speed to 16MHz from 1MHz
Created function init_clocks() to increase clock speed from 1MHz to 16MHz. An important note discovered
is that the clock sources for the MCLK, SMCLK, and ACLK need to be selected to the available clocks BEFORE
disabling fault flags, otherwise the fault flags will continously be set, which resulted in an infinite
loop in this case because the init_clocks() function contains a while loop to try and clear the clock fault
flags before continuing. In this case, ACLK defaults to XT1CLK, so ACLK had to be set to use the REFOCLK
before trying to clear the XT1 fault flag.

Interrupt service routines were added to drivers/io.c to allow interrupts on the available interrupt ports,
which are ports 1 and 2. These will allow external sensors to use interrupts on the IO pins.