feat(uart): Added interrupt control with UART and created UART drivers.
Add a driver for the UART peripheral that can print single characters
to the terminal. Make it interrupt driven and implement it with a
ring buffer for the UART interrupts to get the characters from. This
will be used to implement the printf functionality later.

This was tested by connecting the MSP430F5529 Launchpad to a computer
through the USB cable and using the on-board ez-FET UART to USB module
on the Launchpad. Picocom was used to create a terminal to see values
for UART communication.