feat(printf_assert): Adding printf to asserts
Added printf feature to asserts to print the location of the assert.
Uses mspgcc-elf-addr2line as a separate script to run after an assert
occurs to convert the given program counter at the location of the assert
to a human readable line number in a file. Created a bash script addr2line.sh
in proj/SumoBot/tools/ to run the mspgcc-elf-addr2line instead of having to
run in Makefile; however, it is also included in the Makefile as an option.
Also added the addr2line.sh script to .bashrc and aliased to addr2line so
that it can be run from anywhere.