feat(pwm): Add pwm driver control
Adding pwm_init() to intialize timers connected to PWM pins on Launchpad,
and pwm_set_duty_cycle() to control the duty cycle of the PWM output through
the timer registers via Capture & Compare registers. Have not tested PWM output
on oscilloscope yet prior to this commit, but the TRACE of the PWM output
duty cycle appears to be working.