line_sensor (feat): Added line sensor files to get line crossing.
Created line sensor files as a wrapper for the QRE1113 to get the
corresponding line that is crossed depending on the values read back
from each QRE1113. Tested each ADC input individually with a single
QRE1113 by placing a white piece of paper over the sensor. When the
paper is over the sensor, it returns the enum value for the corresponding
line depending on which input (A1-4) was connected to the QRE1113. If
no paper is over the sensor, the it returned enum value for LINE_NONE.
It appears to be working correctly.